+++ tools/api-markdown-documenter/src/api-item-transforms/default-implementations/TransformApiClass.ts
@@ -14,7 +14,7 @@ import {
 } from "@microsoft/api-extractor-model";
 
 import { SectionNode } from "../../documentation-domain";
-import { ApiModifier, filterByKind, isStatic } from "../ApiItemUtilities";
+import { ApiModifier, filterByKind, isStatic } from "../ApiItemTransformUtilities";
 import { ApiItemTransformationConfiguration } from "../configuration";
 import { createChildDetailsSection, createMemberTables } from "../helpers";
