+++ tools/api-markdown-documenter/src/api-item-transforms/helpers/Helpers.ts
@@ -50,7 +50,7 @@ import {
 	getReturnsBlock,
 	getSeeBlocks,
 	getThrowsBlocks,
-} from "../ApiItemUtilities";
+} from "../ApiItemTransformUtilities";
 import { transformTsdocSection } from "../TsdocNodeTransforms";
 import { getTsdocNodeTransformationOptions } from "../Utilities";
 import { ApiItemTransformationConfiguration } from "../configuration";
